==> notebooks/4.0-ak-building-machine-learning-api.ipynb <==
### Hello World using Flask

import os
hello_world_script_file = os.path.join(os.path.pardir,'src','models','hello_world_api.py')

%%writefile $hello_world_script_file

from flask import Flask, request

app = Flask(__name__)

@app.route('/api', methods=['POST'])
def say_hello():
    data = request.get_json(force=True)
    name = data['name']
    return "hello {0}".format(name)

if __name__ == '__main__':
    app.run(port=10001, debug=True)

import json
import requests

#(py27) kevin@MintXFCE64:~/Projects/titanic/src/models$ pwd
#/home/kevin/Projects/titanic/src/models
#python hello_world_api.py

#need this... python ~/Projects/titanic/src/models/hello_world_api.py

url = 'http://127.0.0.1:10001/api'
data = json.dumps({'name' : 'abhi'})
r = requests.post(url,data)

print r.text

### Machine Learning API using Flask

#### Building API

machine_learning_api_script_file = os.path.join(os.path.pardir,'src','models','machine_learning_api.py')

%%writefile $machine_learning_api_script_file

from flask import Flask, request
import pandas as pd
import numpy as np
import json
import pickle
import os

app = Flask(__name__)

# Load Model and Scaler Files
model_path = os.path.join(os.path.pardir,os.path.pardir,'models')
model_filepath = os.path.join(model_path, 'lr_model.pkl')
scaler_filepath = os.path.join(model_path, 'lr_scaler.pkl')

#todo
model_filepath = '/home/kevin/Projects/titanic/models/lr_model.pkl'
scaler_filepath = '/home/kevin/Projects/titanic/models/lr_scaler.pkl'

scaler = pickle.load(open(scaler_filepath))
model = pickle.load(open(model_filepath))

#todo: take unprocessed input and process it within api
# columns
columns = [ u'Age', u'Fare', u'FamilySize', \
       u'IsMother', u'IsMale', u'Deck_A', u'Deck_B', u'Deck_C', u'Deck_D', \
       u'Deck_E', u'Deck_F', u'Deck_G', u'Deck_Z', u'Pclass_1', u'Pclass_2', \
       u'Pclass_3', u'Title_Lady', u'Title_Master', u'Title_Miss', u'Title_Mr', \
       u'Title_Mrs', u'Title_Officer', u'Title_Sir', u'Fare_Bin_very_low', \
       u'Fare_Bin_low', u'Fare_Bin_high', u'Fare_Bin_very_high', u'Embarked_C', \
       u'Embarked_Q', u'Embarked_S', u'AgeState_Adult', u'AgeState_Child'] 


@app.route('/api', methods=['POST'])
def make_prediction():
    # read json object and conver to json string
    data = json.dumps(request.get_json(force=True))
    # create pandas dataframe using json string
    df = pd.read_json(data)
    # extract passengerIds
    passenger_ids = df['PassengerId'].ravel()
    # actual survived values
    actuals = df['Survived'].ravel()
    # extract required columns based and convert to matrix
    X = df[columns].values.astype('float')
    # transform the input
    X_scaled = scaler.transform(X)
    # make predictions
    predictions = model.predict(X_scaled)
    # create response dataframe
    df_response = pd.DataFrame({'PassengerId': passenger_ids, 'Predicted' : predictions, 'Actual' : actuals})
    # return json 
    return df_response.to_json()

if __name__ == '__main__':
    # host flask app at port 10001
    app.run(port=10001, debug=True)

In order to run the API on the local machine, go to console and run the script.

#need this... python ~/Projects/titanic/src/models/machine_learning_api.py

#### Invoking API using Requests

import os
import pandas as pd
processed_data_path = os.path.join(os.path.pardir,'data','processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
train_df = pd.read_csv(train_file_path)

survived_passengers = train_df[train_df['Survived'] == 1][:5]

survived_passengers

import requests
def make_api_request(data):
    # url for api
    url = 'http://127.0.0.1:10001/api'
    # make post request
    r = requests.post(url,data)
    #return r.text
    return r.json()
    

make_api_request(survived_passengers.to_json())

# compare results
result = make_api_request(train_df.to_json())
df_result = pd.read_json(json.dumps(result))
df_result.head()

# accuracy level
import numpy as np
np.mean(df_result.Actual == df_result.Predicted)
